Removed redundant if clause in getAutoClass()
The clause `else if(extension_loaded('pdo_sqlite') && is_writeable($path))` is redundant, because this code path is only executed, if $path is not writeable. So the right part of that and expression is always false and thus the driver will never be set to sqlite.

Additionally #57 claims, that sqlite is a lot slower than files so there is no point in preferring this driver over files by default.